Add files via upload
The official release! We have Speed, "Flight", Slowfall, Fullbright, Highjump, and Criticals. Currently the mods are potion effects, and that is why this is only available for singleplayer at the moment. Press the righ arrow to open the main GUI!
(flight is a bit broken atm, it is just levitation. we will fix this and make real flight soon.)